split recordings endpoints by active vs archived
